Split ND convolution stuff from conv.py.
